Fixed backend and added constant types for testing

+++ app/Http/Controllers/GenreController.php
@@ -73,8 +73,6 @@ class GenreController extends Controller
             'image' => 'sometimes|image'
         ]);
 
-
-
         if($request->has('title') && $request->title !== null) {
             $genre->title = $request->title;
         }
@@ -88,12 +86,14 @@ class GenreController extends Controller
         }
 
         if($request->hasFile('image')) {
-            Storage::disk('public-images')->delete($genre->image->path);
+            if($genre->image != null) {
+                Storage::disk('public-images')->delete($genre->image->path);
 
+            }
             $path = $request->file('image')->storePublicly(Image::STORING_PATH);
             $path = str_replace(Image::STORING_PATH.'/', '', $path);
 
-            $genre->image->update(['path' => $path]);
+            $genre->image()->updateOrCreate(['path' => $path]);
         }
 
         $genre->save();
